requirements.txt
flask
gunicorn
requests
azure-functions
azure-cosmos